[MIG] Migrated donation_sale v10 to v11
