+++ donation_sale/data/cron.xml
@@ -7,17 +7,16 @@
 
 <odoo noupdate="1"> <!-- noupdate = 1 for the 'active' field -->
 
-<record id="donation_tax_receipt_invoice_cron" model="ir.cron">
-    <field name="name">Generation donation tax receipts from invoices</field>
-    <field name="active" eval="False"/>
-    <field name="user_id" ref="base.user_root"/>
-    <field name="interval_number">1</field>
-    <field name="interval_type">days</field>
-    <field name="numbercall">-1</field> <!-- don't limit the number of calls -->
-    <field name="doall" eval="False"/>
-    <field name="model" eval="'account.invoice'"/>
-    <field name="function" eval="'_generate_each_tax_receipts'" />
-    <field name="args" eval="'()'"/>
-</record>
+    <record id="donation_tax_receipt_invoice_cron" model="ir.cron">
+        <field name="name">Generation donation tax receipts from invoices</field>
+        <field name="active" eval="False"/>
+        <field name="user_id" ref="base.user_root"/>
+        <field name="interval_number">1</field>
+        <field name="interval_type">days</field>
+        <field name="numbercall">-1</field> <!-- don't limit the number of calls -->
+        <field name="doall" eval="False"/>
+        <field name="model_id" ref="model_account_invoice"/>
+        <field name="code">model._generate_each_tax_receipts()</field>
+    </record>
 
 </odoo>
