+++ donation_sale/models/__init__.py
@@ -1,3 +1,6 @@
 # -*- coding: utf-8 -*-
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
+from . import account_invoice
+from . import sale
 from . import donation_sale
