+++ donation_sale/views/donation_tax.xml
@@ -7,19 +7,17 @@
 
 <odoo>
 
-
-<record id="donation_tax_receipt_form" model="ir.ui.view">
-    <field name="name">donation_sale.donation.tax.receipt.form</field>
-    <field name="model">donation.tax.receipt</field>
-    <field name="inherit_id" ref="donation_base.donation_tax_receipt_form"/>
-    <field name="arch"  type="xml">
-        <group name="main" position="after">
-            <group name="invoices" string="Related Invoices">
-                <field name="invoice_ids" nolabel="1"/>
+    <record id="donation_tax_receipt_form" model="ir.ui.view">
+        <field name="name">donation_sale.donation.tax.receipt.form</field>
+        <field name="model">donation.tax.receipt</field>
+        <field name="inherit_id" ref="donation_base.donation_tax_receipt_form"/>
+        <field name="arch"  type="xml">
+            <group name="main" position="after">
+                <group name="invoices" string="Related Invoices">
+                    <field name="invoice_ids" nolabel="1"/>
+                </group>
             </group>
-        </group>
-    </field>
-</record>
-
+        </field>
+    </record>
 
 </odoo>
